Add documentation generation action
This newly created action will trigger on a tagged event, mainly plugin
releases, and will generated the documentation configuration needed for
registering a plugins docs on packer.io.

The action requires access to open a pull-request on the consuming
branch. Once generated and merged the author of the said plugin can open
a pull-request on hashicorp/packer with the details outlined in the
action header for plugins docs to appear under packer.io/docs/...