web-docs: add valid labels
Since the base template in scaffolding contained invalid metadata
labels, this would make new plugins based on it crash when integrated
since the community label does not exist.

This is especially problematic since the failure happens in our systems,
and is not visible directly by the user, so this slows down users'
adoption of our documentations.

So to simplify this and have a ready-to-modify HCL file, we add the two
(for now) supported labels to the metadata.hcl file in order to document
them, so new users can pick it and remove what they don't need from it.